feat(économie-collaborative): introduit le choix entre les 3 régimes de location de meublé
