+++ modele-social/règles/économie-collaborative/location-de-meublé/période.publicodes
@@ -1,0 +1,12 @@
+location de logement meublé . date de début d'activité:
+      par défaut: période . début d'année . N-1
+      question: Quand avez-vous commencé la location de meublé ?
+      type: date
+      suggestions:
+        Début d’année: période . début d'année
+        Début 2024: 01/01/2024
+        Il y a plus de 10 ans: 01/01/2015
+
+location de logement meublé . première année:
+      formule: période . début d'année <= date de début d'activité
+
