+++ modele-social/règles/économie-collaborative/location-de-meublé/régime-général.publicodes
@@ -1,0 +1,34 @@
+location de logement meublé . cotisations . régime général:
+  titre: Cotisations dues pour la location d’un logement meublé de courte durée
+  variations:
+    - si: recettes < seuil de professionalisation
+      alors: 0 €/an
+    - sinon:
+        produit:
+          - assiette
+          - taux de cotisation
+  unité: €/an
+
+  avec:
+    assiette:
+      variations:
+        - si: première année
+          alors: assiette première année
+        - sinon: assiette normale
+      abattement: 60%
+      unité: €/an
+
+      avec:
+        assiette première année:
+          formule: recettes - seuil de professionalisation
+          plancher: 0
+          unité: €/an
+        assiette normale: recettes
+
+    taux de cotisation:
+      titre: Taux de cotisation au régime général
+      description: Taux de cotisation pour la location d’un logement meublé de courte durée au régime général
+      variations:
+        - si: zone géographique . Alsace-Moselle
+          alors: 48.72%
+        - sinon:  47.42%
