+++ modele-social/règles/économie-collaborative/location-de-meublé/zone-géographique.publicodes
@@ -1,0 +1,15 @@
+location de logement meublé . zone géographique:
+  question: Quelle est votre zone géographique ?
+  une possibilité:
+    possibilités:
+      - Alsace-Moselle
+      - autre
+    choix obligatoire: oui
+  par défaut: "'autre'"
+  avec:
+    Alsace-Moselle:
+      titre: Bas-Rhin, Haut-Rhin ou Moselle
+      valeur: zone géographique = 'Alsace-Moselle'
+    autre:
+      titre: Autre département
+      valeur: zone géographique = 'autre'
