+++ site/source/components/conversation/OuiNonInput.tsx
@@ -1,0 +1,39 @@
+import { Trans, useTranslation } from 'react-i18next'
+
+import { InputProps } from '@/components/conversation/RuleInput'
+import { Radio, ToggleGroup } from '@/design-system'
+import { useSelection } from '@/hooks/UseSelection'
+
+export function OuiNonInput(props: InputProps) {
+	const { t } = useTranslation()
+
+	// seront stockées ainsi dans le state :
+	// [parent object path]: dotted fieldName relative to parent
+	const { handleChange, defaultValue, currentSelection } = useSelection(props)
+
+	return (
+		<ToggleGroup
+			aria-label={t('conversation.yes-no.aria-label', 'Oui ou non')}
+			{...props}
+			onChange={handleChange}
+			value={currentSelection ?? undefined}
+		>
+			{/* eslint-disable-next-line jsx-a11y/no-autofocus */}
+			<Radio
+				value="oui"
+				id={`input-oui-${props.id || ''}`}
+				autoFocus={props.autoFocus && defaultValue === 'oui'}
+			>
+				<Trans>Oui</Trans>
+			</Radio>
+			{/* eslint-disable-next-line jsx-a11y/no-autofocus */}
+			<Radio
+				value="non"
+				id={`input-non-${props.id || ''}`}
+				autoFocus={props.autoFocus && defaultValue === 'non'}
+			>
+				<Trans>Non</Trans>
+			</Radio>
+		</ToggleGroup>
+	)
+}
