+++ site/source/components/conversation/RuleInput.tsx
@@ -9,19 +9,21 @@ import Engine, {
 } from 'publicodes'
 import React from 'react'
 
-import NumberInput from '@/components/conversation/NumberInput'
-import SelectCommune from '@/components/conversation/select/SelectCommune'
 import { useEngine } from '@/components/utils/EngineContext'
 import { DateFieldProps } from '@/design-system/field/DateField'
 import { Spacing } from '@/design-system/layout'
 import { getMeta } from '@/utils/publicodes'
 
 import { normalizeRuleName } from '../utils/normalizeRuleName'
-import { Choice, MultipleAnswerInput, OuiNonInput } from './ChoicesInput'
+import { Choice } from './Choice'
 import DateInput from './DateInput'
 import { DefaultValue } from './DefaultValue'
 import { MultipleChoicesInput } from './MulipleChoicesInput'
+import { MultipleAnswerInput } from './MultipleAnswerInput'
+import NumberInput from './NumberInput'
+import { OuiNonInput } from './OuiNonInput'
 import ParagrapheInput from './ParagrapheInput'
+import SelectCommune from './select/SelectCommune'
 import SelectPaysDétachement from './select/SelectPaysDétachement'
 import SelectAtmp from './select/SelectTauxRisque'
 import TextInput from './TextInput'
