+++ site/source/domaine/relativeDottedName.ts
@@ -1,0 +1,4 @@
+export const relativeDottedName = (
+	rootDottedName: string,
+	childDottedName: string
+) => childDottedName.replace(rootDottedName + ' . ', '')
