+++ site/source/pages/simulateurs/comparaison-statuts/components/ModifierOptions.tsx
@@ -2,9 +2,9 @@ import { PublicodesExpression } from 'publicodes'
 import { Trans, useTranslation } from 'react-i18next'
 import { styled } from 'styled-components'
 
-import { SwitchInput } from '@/components/conversation/ChoicesInput'
 import { ExplicableRule } from '@/components/conversation/Explicable'
 import RuleInput from '@/components/conversation/RuleInput'
+import { SwitchInput } from '@/components/conversation/SwitchInput'
 import { Message } from '@/design-system'
 import { Button } from '@/design-system/buttons'
 import { Drawer } from '@/design-system/drawer'
