+++ site/source/pages/simulateurs/location-de-meublé/simulationConfig.ts
@@ -3,5 +3,10 @@ import { SimulationConfig } from '@/domaine/SimulationConfig'
 export const configLocationDeMeublé: SimulationConfig = {
 	objectifs: ['location de logement meublé . cotisations'],
 	'unité par défaut': '€/an',
-	situation: {},
+	questions: {
+		liste: ['location de logement meublé . affiliation', ''],
+	},
+	situation: {
+		'entreprise . activité . nature': "'commerciale'",
+	},
 }
